Move _get_annotated_movie into test_utils.py and ignore from coverage.
