+++ .coveragerc
@@ -14,3 +14,4 @@ show_missing = True
 
 omit =
     **/*_test.py
+    deepcell_tracking/test_utils.py
